合并工程

+++ circularBuffer2.c
@@ -1,10 +1,12 @@
+#include <stdio.h>
+
 typedef struct circular_buffer
 {
 	void *buffer;     // data buffer
 	void *buffer_end; // end of data buffer
-	size_t capacity;  // maximum number of items in the buffer
-	size_t count;     // number of items in the buffer
-	size_t sz;        // size of each item in the buffer
+	unsigned int capacity;  // maximum number of items in the buffer
+	unsigned int count;     // number of items in the buffer
+	unsigned int sz;        // size of each item in the buffer
 	void *head;       // pointer to head
 	void *tail;       // pointer to tail
 } circular_buffer;
